Tidy emp_assert, revealing emp_always_assert
