+++ include/emp/base/_tdebug_assert_trigger.hpp
@@ -1,0 +1,41 @@
+/**
+ *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
+ *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
+ *  @date 2020.
+ *
+ *  @file _tdebug_assert_trigger.hpp
+ *  @brief Non-terminating asssert trigger implementation for unit testing.
+ *  @note For internal use.
+ */
+
+#ifndef EMP_TDEBUG_ASSERT_TRIGGER_HPP
+#define EMP_TDEBUG_ASSERT_TRIGGER_HPP
+
+namespace emp {
+
+  constexpr bool assert_on = true;
+  struct AssertFailInfo {
+    std::string filename;
+    int line_num;
+    std::string error;
+  };
+  AssertFailInfo assert_fail_info;
+  bool assert_last_fail = false;
+
+  template <typename... EXTRA>
+  bool assert_trigger(
+    std::string filename, size_t line, std::string expr, EXTRA &&... extra
+  ) {
+    emp::assert_fail_info.filename = __FILE__;
+    emp::assert_fail_info.line_num = __LINE__;
+    emp::assert_fail_info.error = expr;
+    emp::assert_last_fail = true;
+
+    return false; // do not process subsequent abort
+  }
+
+  void assert_clear() { emp::assert_last_fail = false; }
+
+} // namespace emp
+
+#endif // #ifndef EMP_TDEBUG_ASSERT_TRIGGER_HPP
